pyproject.toml
[project]
name = "alienworlds"
version = "0.1.0"
description = "Alien Worlds Pool Monitor"
requires-python = ">=3.13"
dependencies = [
    "prometheus-client>=0.20.0",
    "eosapi-async>=2.0.1",
    "colorlog>=6.8.0",
    "cryptos>=2.0.9",
    "base58>=2.1.1",
    "cachetools>=5.5.1",
    "pydantic>=2.10.6",
    "antelopy>=0.2.0",
    "psycopg>=3.2.4",
    "psycopg-pool>=3.2.4",
    "python-dotenv>=1.0.1",
    "cryptography>=44.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["alienworlds"]
